+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 7.0.0
 
+### Fixes
+
 - Fix: Remove `SentryOptions` related parameters from classes which also take `Hub` as a parameter (#816)
 
 ### Breaking Changes
@@ -13,6 +15,7 @@
     - screenResolution
     - theme
   - Removed isolate name from Dart context. It's now reported via the threads interface. It can be enabled via `options.attachThreads`
+- Use `sentryClientName` instead of `sdk.identifier` ([#1135](https://github.com/getsentry/sentry-dart/pull/1135))
 
 ## 6.11.1