+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 7.0.0
 
+### Features
+
+- Feat: Introduce Hint data bag ([#1136](https://github.com/getsentry/sentry-dart/pull/1136))
+
+### Fixes
+
 - Fix: Remove `SentryOptions` related parameters from classes which also take `Hub` as a parameter (#816)
 
 ### Breaking Changes